Implemented async image caching, refactored utils into their own file
